Chore: removed bitcoinjs-lib dependent code.
