+++ tests/stacks.test.ts
@@ -14,12 +14,12 @@ describe('suite', () => {
 
   it.concurrent('stacks: isAllowed() check login allowed from known address', async () => {
     const res = isAllowed(allowed[0].stx);
-    expect(res).equals(true);
+    expect(typeof res).equals('object');
   })
 
   it.concurrent('stacks: isAllowed() check login not allowed from unknown address', async () => {
     const res = isAllowed(allowed[0].stx + '1');
-    expect(res).equals(false);
+    expect(typeof res).equals('undefined');
   })
 
   it.concurrent('stacks: decodeStacksAddress() check mainnet stacks address can be decoded', async () => {
